Services/AM: Expose CIAFile class in AM header
